backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.9.0
spacy>=3.8.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
